Add extension to Date for formatting and converting Date to String
