+++ TOZ_iOS/Date+Formatter.swift
@@ -1,0 +1,17 @@
+//
+//  Date+Formatter.swift
+//  TOZ_iOS
+//
+//  Copyright © 2017 intive. All rights reserved.
+//
+
+import Foundation
+
+extension Date {
+
+    func dateToFormattedString() -> String {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "yyyy-MM-dd"
+        return formatter.string(from: self)
+    }
+}
