+++ TOZ_iOS/NewsDetailViewController.swift
@@ -31,12 +31,3 @@ class NewsDetailViewController: UIViewController {
     }
 
 }
-
-extension Date {
-
-    func dateToFormattedString() -> String {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd"
-        return formatter.string(from: self)
-    }
-}
